+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/hooks/AuthMeHook.java
@@ -6,7 +6,10 @@ import org.bukkit.entity.Player;
 
 /**
  * Github: https://github.com/Xephi/AuthMeReloaded/
- * Project page: http://dev.bukkit.org/bukkit-plugins/authme-reloaded/
+ * Project page:
+ *
+ * Bukkit: http://dev.bukkit.org/bukkit-plugins/authme-reloaded/
+ * Spigot: https://www.spigotmc.org/resources/authme-reloaded.6269/
  */
 public class AuthMeHook implements AuthPlugin {