+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/hooks/AuthPlugin.java
@@ -8,7 +8,8 @@ import org.bukkit.entity.Player;
 public interface AuthPlugin {
 
     /**
-     * Login the premium (paid account) player
+     * Login the premium (paid account) player after
+     * the player joined successfully the server.
      *
      * @param player the player that needs to be logged in
      */
@@ -21,6 +22,9 @@ public interface AuthPlugin {
      * so we can be sure the premium player doesn't steal the account
      * of that player.
      *
+     * This operation will be performed async while the player is
+     * connecting
+     *
      * @param playerName player name
      * @return if the player has an account
      */
@@ -28,6 +32,10 @@ public interface AuthPlugin {
 
     /**
      * Forces a register in order to protect the paid account.
+     * The method will be invoked after the player joined the server.
+     *
+     * After a successful registration the player should be logged
+     * in too.
      *
      * The method will be called only for premium accounts.
      * So it's recommended to set additionally premium property